+++ src/utils/helper.js
@@ -27,127 +27,126 @@ async function generatePrivData(mnemonic, pin) {
   return priv;
 }
 
-
-async function removeEmptyAccounts(indexAddress, keyringInstance, vaultState, unmarshalApiKey, recoverMechanism, logs) {
+async function getAccountsFromTransactions(indexAddress, keyringInstance, vaultState, unmarshalApiKey) {
   
   const keyring = keyringInstance.getKeyringsByType(vaultState.keyrings[0].type);
 
   let zeroCounter = 0;
-  let accountCheckList = [];
   let accountsArray = [];
-  accountsArray.push({ address: indexAddress, isDeleted: false, isImported: false, label: 'Wallet 1' });
-  accountCheckList.push(indexAddress)
-  let labelCounter = 2;  // as an initial wallet is already created above with label 'Wallet 1'
-  const chains = Object.keys(Chains.evmChains);
+  accountsArray.push({ address: indexAddress, isDeleted: false, isImported: false, label: 'EVM Wallet 1' });
 
-  let newAccountAddr = indexAddress
+  do {
+    zeroCounter = 0;
+    for(let i=0; i < 5; i++) {
+      const vaultState = await keyringInstance.addNewAccount(keyring[0]);
 
-  if( recoverMechanism === 'logs'){
-    for(let i=0; i < logs.length; i++){
-      if (logs[i].action === 'add-account' && (chains.includes(logs[i].chain) || logs[i].chain === undefined)){
-        if (accountCheckList.includes(newAccountAddr)) {
-          vaultState = await keyringInstance.addNewAccount(keyring[0]);
-          newAccountAddr = Web3.utils.toChecksumAddress(vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1])
-        }
-        
-        if (logs[i].address.toLowerCase() !== newAccountAddr.toLowerCase() && !accountCheckList.includes(logs[i].address.toLowerCase())) {
-          do {
-            const label = this.createWalletLabels('all', labelCounter);
-            accountsArray.push({ address: newAccountAddr.toLowerCase(), isDeleted: false, isImported: false, label, isExported: false });
-            accountCheckList.push(newAccountAddr.toLowerCase())
-            labelCounter++;
-            let vaultState = await keyringInstance.addNewAccount(keyring[0]);
-            newAccountAddr = Web3.utils.toChecksumAddress(vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1])
-          }
-          while(logs[i].address.toLowerCase() !== newAccountAddr.toLowerCase() && !accountCheckList.includes(logs[i].address.toLowerCase()))
-        }
-        
-        if (logs[i].address.toLowerCase() === newAccountAddr.toLowerCase()) {
-          const label = this.createWalletLabels('all', labelCounter);
-          accountsArray.push({ address: newAccountAddr.toLowerCase(), isDeleted: false, isImported: false, label, isExported: false });
-          accountCheckList.push(newAccountAddr.toLowerCase())
-          labelCounter++;
-        }
-        
-      }
-      if(logs[i].action === 'delete-account' && (chains.includes(logs[i].chain) || logs[i].chain === undefined)) {
-        let ind = accountsArray.findIndex((acc) => acc.address === Web3.utils.toChecksumAddress(logs[i].address))
-        ind >= 0 ? accountsArray[ind].isDeleted = true : false;
+      const ethActivity = await getETHTransactions(vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1], 'ethereum', unmarshalApiKey);
+      const polygonActivity = await getPolygonTransactions(vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1], 'polygon', unmarshalApiKey);
+      const bscActivity = await getBSCTransactions(vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1], 'bsc', unmarshalApiKey);
+      const label = this.createWalletLabels('EVM', i+2);
+
+      if (!ethActivity && !polygonActivity && !bscActivity) {
+        accountsArray.push({ address: vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1], isDeleted: true, isImported: false, label, isExported: false });
+        zeroCounter++;
+      } else {
+        accountsArray.push({ address: vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1], isDeleted: false, isImported: false, label, isExported: false });
+        zeroCounter = 0;
       }
     }
-
-  } else if( recoverMechanism === 'transactions'){
-    do {
-      zeroCounter = 0;
-      for(let i=0; i < 5; i++) {
-        const vaultState = await keyringInstance.addNewAccount(keyring[0]);
-  
-        const ethActivity = await getETHTransactions(vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1], 'ethereum', unmarshalApiKey);
-        const polygonActivity = await getPolygonTransactions(vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1], 'polygon', unmarshalApiKey);
-        const bscActivity = await getBSCTransactions(vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1], 'bsc', unmarshalApiKey);
-        const label = this.createWalletLabels('all', i+2);
-  
-        if (!ethActivity && !polygonActivity && !bscActivity) {
-          accountsArray.push({ address: vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1], isDeleted: true, isImported: false, label, isExported: false });
-          zeroCounter++;
-        } else {
-          accountsArray.push({ address: vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1], isDeleted: false, isImported: false, label, isExported: false });
-          zeroCounter = 0;
-        }
-      }
-    }
-  
-    while (zeroCounter < 5 )
   }
+
+  while (zeroCounter < 5 )
+
   return accountsArray;
 }
 
-async function  getAccountsFromLogs(keyringInstance, vaultState, recoverMechanism, logs) {
-
-  //if mech = transaction - generate one acc for bitcoin
+async function  getAccountsFromLogs(chain, chainInstance, vaultState, logs, indexAddress) {
   
   let accountsArray = [];
-  let accountCheckList = [];
-  let {address} = await keyringInstance.addAccount();
-  const label = this.createWalletLabels('bitcoin', 1);
-  accountsArray.push({ address: address, isDeleted: false, isImported: false, label, isExported: false });
-  accountCheckList.push(address)
-  let labelCounter = 2;
-  const chains = Object.keys(Chains.nonEvmChains);
+  const accountCheckSet = new Set();
+  const evmChains = Object.keys(Chains.evmChains);
+  let generatedAddress = indexAddress
+  let labelCounter = 1;
+  let keyring, labelPrifix
 
-  if( recoverMechanism === 'logs'){
-    for(let i=0; i < logs.length; i++){
-      if (logs[i].action === 'add-account' && (chains.includes(logs[i].chain))){
+  if (chain === 'ethereum' || chain === undefined || evmChains.includes(chain)) {
+    keyring = chainInstance.getKeyringsByType(vaultState.keyrings[0].type);
+    labelPrifix = 'EVM';
+  } else {
+    labelPrifix = Chains.nonEvmChains[chain];
+  }
 
-        if (accountCheckList.includes(address)) {
-          address = (await keyringInstance.addAccount()).address;
+  // create account sequentially using chain specific keyring instance
+  const createNewAddress = async (chain, chainInstance) => {
+    let address
+    if (chain === 'ethereum' || chain === undefined) {
+      vaultState = await chainInstance.addNewAccount(keyring[0])
+      address = (Web3.utils.toChecksumAddress(vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1])).toLowerCase()
+    }
+    else {
+      address = (await chainInstance.addAccount()).address.toLowerCase();
+    }
+    return address;
+  }
+
+  // create account data wrt vault structure
+  const createAccountObject = async (generatedAddress) => {
+    const label = this.createWalletLabels(labelPrifix, labelCounter++);
+    return { address: generatedAddress, isDeleted: false, isImported: false, label, isExported: false };
+  };
+
+  // add the initial address in the account array
+  if (!indexAddress){
+    return accountsArray
+  }
+  else if(!logs) {
+    const account = await createAccountObject(indexAddress);
+    accountsArray.push(account);
+    accountCheckSet.add(account.address.toLowerCase());
+    return accountsArray
+  }
+  else {
+    const account = await createAccountObject(indexAddress);
+    accountsArray.push(account);
+    accountCheckSet.add(account.address.toLowerCase());
+  }
+
+
+  const addAccountIfVerified = async (logAddress) => {
+
+    if (accountCheckSet.has(generatedAddress)) {
+      // If the previous generated address was added in the check set, generate a new address
+      generatedAddress = await createNewAddress(chain, chainInstance);
+    }
+  
+    while (logAddress !== generatedAddress && !accountCheckSet.has(logAddress)) {
+      // Generate new addresses till the one matching the logAddress, which has not been addded to checkset yet
+      const account = await createAccountObject(generatedAddress);
+      accountsArray.push(account);
+      accountCheckSet.add(account.address.toLowerCase());
+      generatedAddress = await createNewAddress(chain, chainInstance);
+    }
+  
+    if (logAddress === generatedAddress) {
+      // If the address matches, add it to the accounts array
+      const account = await createAccountObject(generatedAddress);
+      accountsArray.push(account);
+      accountCheckSet.add(account.address.toLowerCase());
+    }
+  };
+
+  for (let log of logs) {
+    const logAddress = log?.address?.toLowerCase();
+    if (log.action === 'add-account' && log?.chain === chain) {
+      await addAccountIfVerified(logAddress);
+    } else if (log.action === 'delete-account' && Chains.nonEvmChains.hasOwnProperty(log.chain)) {
+        const index = accountsArray.findIndex((acc) => acc.address === logAddress);
+        if (index !== -1) {
+          accountsArray[index].isDeleted = true;
         }
-        
-        if (logs[i].address.toLowerCase() !== address.toLowerCase() && !accountCheckList.includes(logs[i].address.toLowerCase())) {
-          do {
-            const label = this.createWalletLabels('bitcoin', labelCounter);
-            accountsArray.push({ address: address.toLowerCase(), isDeleted: false, isImported: false, label, isExported: false });
-            accountCheckList.push(address.toLowerCase())
-            labelCounter++;
-            address  = (await keyringInstance.addAccount()).address;
-          }
-          while(logs[i].address.toLowerCase() !== address.toLowerCase() && !accountCheckList.includes(logs[i].address.toLowerCase()))
-        }
-        
-        if (logs[i].address.toLowerCase() === address.toLowerCase()) {
-          const label = this.createWalletLabels('bitcoin', labelCounter);
-          accountsArray.push({ address: address.toLowerCase(), isDeleted: false, isImported: false, label, isExported: false });
-          accountCheckList.push(address.toLowerCase())
-          labelCounter ++;
-        }
-        
-      }
-      if(logs[i].action === 'delete-account' && (chains.includes(logs[i].chain))) {
-        let ind = accountsArray.findIndex((acc) => acc.address.toLowerCase() === logs[i].address.toLowerCase())
-        ind >= 0 ? accountsArray[ind].isDeleted = true : false;
-      }
     }
   }
+
   return accountsArray;
 }
 
@@ -284,32 +283,16 @@ function validateEncryptionKey(data, encryptionKey, encryptor, isCustomEncryptor
   }
 
 
-function createWalletLabels(labelObj = 'all', walletIndex = 1) {
-  let labels = {};
-
-  const chains = Object.keys(Chains.evmChains);
-  
-  if (labelObj === 'all') {
-    chains.forEach(chain => labels[chain] = `${chain.charAt(0).toUpperCase() + chain.substr(1).toLowerCase()} Wallet ${walletIndex}` );
-  } 
-  else if (labelObj === 'bitcoin') {
-    labels = `${labelObj.charAt(0).toUpperCase() + labelObj.substr(1).toLowerCase()} Wallet ${walletIndex}`;
-  }
-  else {
-    chains.forEach(chain => {
-      if (labels[chain] !== undefined) {
-        labels[chain] = `${chain.charAt(0).toUpperCase() + chain.substr(1).toLowerCase()} Wallet ${walletIndex}`;
-      }
-    })
-  }
-
+function createWalletLabels(labelPrefix, walletIndex = 1) {
+  let labels = `${labelPrefix} Wallet ${walletIndex}`;
   return labels;
 }
 
 module.exports = {
   stringToArrayBuffer,
   generatePrivData,
-  removeEmptyAccounts,
+  // removeEmptyAccounts,
+  getAccountsFromTransactions,
   getAccountsFromLogs,
   getCoinInstance,
   getAssetDetails,